refactor: split config and CLI
