+++ src/api/accounts/stake_address/root.rs
@@ -2,7 +2,7 @@ use crate::{
     BlockfrostError,
     accounts::{AccountData, AccountsPath},
     api::ApiResult,
-    cli::Config,
+    config::Config,
 };
 
 use axum::{Extension, extract::Path};
